Moved all error definitions in //ledger/error
