+++ ledger/error/src/main/scala/com/daml/error/ErrorCause.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.store
+package com.daml.error
 
 import com.daml.lf.engine.{Error => LfError}
