+++ ledger/error/src/main/scala/com/daml/error/definitions/ErrorGroups.scala
@@ -1,7 +1,4 @@
-// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.platform.apiserver.error
+package com.daml.error.definitions
 
 import com.daml.error.{ErrorClass, ErrorGroup}
