+++ ledger/error/src/main/scala/com/daml/error/definitions/RejectionGenerators.scala
@@ -1,18 +1,14 @@
-// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
+package com.daml.error.definitions
 
-package com.daml.platform.apiserver.error
-
-import com.daml.error.{BaseError, ErrorCodeLoggingContext}
+import com.daml.error.{BaseError, ErrorCause, ErrorCodeLoggingContext}
 import com.daml.ledger.participant.state
 import com.daml.lf.engine.Error.{Interpretation, Package, Preprocessing, Validation}
-import com.daml.lf.engine.{Error => LfError}
-import com.daml.lf.interpretation.{Error => LfInterpretationError}
-import com.daml.platform.store.ErrorCause
 import io.grpc.Status.Code
 import io.grpc.StatusRuntimeException
 
 import scala.util.{Failure, Success, Try}
+import com.daml.lf.engine.{Error => LfError}
+import com.daml.lf.interpretation.{Error => LfInterpretationError}
 
 object RejectionGenerators {
