+++ ledger/participant-integration-api/BUILD.bazel
@@ -39,7 +39,6 @@ compile_deps = [
     "//daml-lf/transaction",
     "//daml-lf/transaction:transaction_proto_java",
     "//daml-lf/transaction:value_proto_java",
-    "//daml-lf/validation",
     "//language-support/scala/bindings",
     "//ledger-api/rs-grpc-akka",
     "//ledger-api/rs-grpc-bridge",
@@ -87,7 +86,6 @@ compile_deps = [
 
 scala_compile_deps = [
     "@maven//:com_github_scopt_scopt",
-    "@maven//:org_typelevel_cats_core",
     "@maven//:com_typesafe_akka_akka_actor",
     "@maven//:com_typesafe_akka_akka_stream",
     "@maven//:org_playframework_anorm_anorm",
