+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/execution/StoreBackedCommandExecutor.scala
@@ -3,30 +3,21 @@
 
 package com.daml.platform.apiserver.execution
 
+import com.daml.error.ErrorCause
+
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicLong
-
 import com.daml.ledger.api.domain.{Commands => ApiCommands}
 import com.daml.ledger.configuration.Configuration
 import com.daml.ledger.participant.state.index.v2.{ContractStore, IndexPackagesService}
 import com.daml.ledger.participant.state.{v2 => state}
 import com.daml.lf.crypto
 import com.daml.lf.data.{ImmArray, Ref}
-import com.daml.lf.engine.{
-  Engine,
-  Result,
-  ResultDone,
-  ResultError,
-  ResultNeedContract,
-  ResultNeedKey,
-  ResultNeedPackage,
-  Error => DamlLfError,
-}
+import com.daml.lf.engine.{Engine, Result, ResultDone, ResultError, ResultNeedContract, ResultNeedKey, ResultNeedPackage, Error => DamlLfError}
 import com.daml.lf.transaction.Node
 import com.daml.logging.LoggingContext
 import com.daml.metrics.{Metrics, Timed}
 import com.daml.platform.packages.DeduplicatingPackageLoader
-import com.daml.platform.store.ErrorCause
 import scalaz.syntax.tag._
 
 import scala.concurrent.{ExecutionContext, Future}
