+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/execution/TimedCommandExecutor.scala
@@ -3,12 +3,12 @@
 
 package com.daml.platform.apiserver.execution
 
+import com.daml.error.ErrorCause
 import com.daml.ledger.api.domain
 import com.daml.ledger.configuration.Configuration
 import com.daml.lf.crypto
 import com.daml.logging.LoggingContext
 import com.daml.metrics.{Metrics, Timed}
-import com.daml.platform.store.ErrorCause
 
 import scala.concurrent.{ExecutionContext, Future}
